# landing-page/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SalonHub API
  version: "1.0.0"
  description: |
    SalonHub is a multi-tenant SaaS platform designed for beauty salons, restaurants,
    training centers, and medical practices. The API provides comprehensive management
    of appointments, clients, services, staff, settings, and public booking flows.

    ## Authentication

    The API supports two authentication methods:

    - **JWT Bearer Token** -- Obtained via `/auth/login` or `/auth/register`. Include as `Authorization: Bearer <token>`.
    - **API Key** -- Prefixed with `sk_live_`. Include as `Authorization: Bearer sk_live_...`. API keys are scoped and can be managed via the `/api-keys` endpoints.

    ## Multi-Tenancy

    Most endpoints require tenant context. The tenant is resolved automatically from the
    authenticated user's JWT token. Public endpoints use the tenant slug in the URL path.

    ## Error Handling

    All error responses follow a consistent format:
    ```json
    {
      "success": false,
      "error": "ERROR_CODE",
      "message": "Human-readable description"
    }
    ```
  contact:
    name: SalonHub Support
    email: support@salonhub.app
    url: https://salonhub.app
  license:
    name: Proprietary
    url: https://salonhub.app/terms
servers:
  - url: https://api.salonhub.app/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api-keys:
    post:
      tags: [API Keys]
      summary: Create an API key
      description: |
        Creates a new API key for the current tenant. Requires owner or admin role
        and a developer, custom, or trial subscription plan.

        **Important:** The full API key (prefixed with `sk_live_`) is returned only once
        in this response. Store it securely.
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
            example:
              name: "Mobile App Integration"
              scopes: ["clients:read", "appointments:read", "appointments:write"]
              expires_at: "2027-03-01T00:00:00.000Z"
      responses:
        "201":
          description: API key created (full key returned only once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ApiKeyFull"
              example:
                success: true
                data:
                  id: 1
                  name: "Mobile App Integration"
                  key: "sk_live_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"
                  prefix: "sk_live_a1b2c3d4"
                  scopes: ["clients:read", "appointments:read", "appointments:write"]
                  is_active: true
                  expires_at: "2027-03-01T00:00:00.000Z"
                  created_at: "2026-03-02T12:00:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Insufficient subscription plan or role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                error: "FORBIDDEN"
                message: "API keys require a developer, custom, or trial subscription plan"
    get:
      tags: [API Keys]
      summary: List API keys
      description: Returns all API keys for the current tenant. Only the prefix is shown, not the full key.
      operationId: listApiKeys
      responses:
        "200":
          description: API keys retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
              example:
                success: true
                data:
                  - id: 1
                    name: "Mobile App Integration"
                    prefix: "sk_live_a1b2c3d4"
                    scopes: ["clients:read", "appointments:read", "appointments:write"]
                    is_active: true
                    last_used_at: "2026-03-01T18:45:00.000Z"
                    expires_at: "2027-03-01T00:00:00.000Z"
                    created_at: "2026-03-02T12:00:00.000Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api-keys/{id}:
    patch:
      tags: [API Keys]
      summary: Update an API key
      description: Updates an API key's name, scopes, or active status.
      operationId: updateApiKey
      parameters:
        - $ref: "#/components/parameters/ApiKeyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApiKeyRequest"
            example:
              name: "Mobile App Integration v2"
              scopes: ["clients:read", "clients:write", "appointments:read", "appointments:write"]
              is_active: true
      responses:
        "200":
          description: API key updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ApiKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [API Keys]
      summary: Delete an API key
      description: Soft-deletes an API key. The key will no longer be usable for authentication.
      operationId: deleteApiKey
      parameters:
        - $ref: "#/components/parameters/ApiKeyId"
      responses:
        "200":
          description: API key deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "API key deleted successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

# ════════════════════════════════════════════════
# COMPONENTS
# ════════════════════════════════════════════════
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/auth/login` or `/auth/register`.
        Include as: `Authorization: Bearer <jwt_token>`
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key prefixed with `sk_live_`. Obtained from `/api-keys`.
        Include as: `Authorization: Bearer sk_live_...`
        Requires developer, custom, or trial subscription plan.

  # ──────────────────────────────────────────────
  # PARAMETERS
  # ──────────────────────────────────────────────
  parameters:
...
    ApiKeyId:
      name: id
      in: path
      required: true
      description: API key ID
      schema:
        type: integer
        example: 1
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "BAD_REQUEST"
            message: "Validation failed: email is required"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "UNAUTHORIZED"
            message: "Authentication required"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "FORBIDDEN"
            message: "You do not have permission to perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "NOT_FOUND"
            message: "Resource not found"

  # ──────────────────────────────────────────────
  # SCHEMAS
  # ──────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [success, error, message]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Machine-readable error code
          example: "BAD_REQUEST"
        message:
          type: string
          description: Human-readable error description
          example: "Validation failed"

    # ── Pagination ──
...
    ApiKey:
      type: object
      description: API key (prefix only, full key is not stored)
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Mobile App Integration"
        prefix:
          type: string
          description: First 16 characters of the key
          example: "sk_live_a1b2c3d4"
        scopes:
          type: array
          items:
            type: string
          example: ["clients:read", "appointments:read", "appointments:write"]
        is_active:
          type: boolean
          example: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-01T18:45:00.000Z"
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: "2027-03-01T00:00:00.000Z"
        created_at:
          type: string
          format: date-time
          example: "2026-03-02T12:00:00.000Z"
    ApiKeyFull:
      type: object
      description: API key with full key value (returned only at creation)
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Mobile App Integration"
        key:
          type: string
          description: The full API key. Only returned once at creation time.
          example: "sk_live_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"
        prefix:
          type: string
          example: "sk_live_a1b2c3d4"
        scopes:
          type: array
          items:
            type: string
          example: ["clients:read", "appointments:read", "appointments:write"]
        is_active:
          type: boolean
          example: true
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: "2027-03-01T00:00:00.000Z"
        created_at:
          type: string
          format: date-time
          example: "2026-03-02T12:00:00.000Z"

    # ── Settings ──
...
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Mobile App Integration"
        scopes:
          type: array
          items:
            type: string
          example: ["clients:read", "appointments:read", "appointments:write"]
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: "2027-03-01T00:00:00.000Z"
    UpdateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          example: "Mobile App Integration v2"
        scopes:
          type: array
          items:
            type: string
          example: ["clients:read", "clients:write", "appointments:read", "appointments:write"]
        is_active:
          type: boolean
          example: true